Index.html
added a seperat folder for pictures.

added the index page. Not done quite yet. I've broken the "fluid" motion of it. need to figure out where i did that. Also the fb logo doesn't want to show up for some reason. need to look into that too. Just using the "Style.css" file to do the css work in. have tried to seperate the two pages though. don't really like the colors either. but we'll flush that out i hope.