Resizing the update
Not quite done yet.

Did add a wepage called Error for the links that havn't got an adress yet.

Still not done with the resizing problems on index. But i'll use my desktop to figure it out on. a lot easier to code on (bigger screen). but most of it has been delt with. Just messing arround with the copyright bar atm.

Removed FB from the links tab, and added patreon instead. (more relevant) Need to figure out how to do the logo's inline though.